changed from using Float to Double (in Config::_scrollHealthDamage) this was due to the exception "java.lang.ClassCastException: class java.lang.Double cannot be cast to class java.lang.Float" the docs does not say anything about this being possible, so might as well just use double and convert it back to float. (it will be the same anyway.)

+++ src/main/java/jeff/iss_addons/config/Config.java
@@ -11,7 +11,7 @@ public class Config
     public final ModConfigSpec.ConfigValue<Boolean> _enableCinderEssenceRecipe;
     public final ModConfigSpec.ConfigValue<Boolean> _enableCommonInkEssenceRecipe;
     public final ModConfigSpec.ConfigValue<Boolean> _enableScrollMod;
-    public final ModConfigSpec.ConfigValue<List<Float>> _scrollHealthDamage;
+    public final ModConfigSpec.ConfigValue<List<Double>> _scrollHealthDamage;
     public final ModConfigSpec.ConfigValue<Boolean> _enableTelekinesisMod;
     public final ModConfigSpec.ConfigValue<Boolean> _enableCounterspellMod;
 
@@ -26,7 +26,7 @@ public class Config
         builder.push("Scroll");
         _enableScrollMod = builder.define("Enable", true);
         builder.comment("damage to player (in hearts) per scroll use, ordered from common to legendary.");
-        _scrollHealthDamage = builder.define("Damage", Arrays.asList(1.0f, 1.5f, 2.0f, 2.5f, 3.0f), new EnsureListSameSize<>(5));
+        _scrollHealthDamage = builder.define("Damage", Arrays.asList(1.0, 1.5, 2.0, 2.5, 3.0), new EnsureListSameSize<>(5));
         builder.pop();
 
         builder.push("Telekinesis");
